perf evsel: Fix use of inherit in non mmap case
perf stat doesn't mmap and its perfectly fine for it to use task-bound
counters with inheritance.

So don't do any change in the perf_evsel__open parameters, leaving the
syscall itself to validate this.

When the mmap fails perf_evlist__mmap will just emit a warning if this
is the failure reason.

Reported-by: Peter Zijlstra <peterz@infradead.org>
Cc: Frederic Weisbecker <fweisbec@gmail.com>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Mike Galbraith <efault@gmx.de>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Stephane Eranian <eranian@google.com>
Cc: Tom Zanussi <tzanussi@gmail.com>
Link: http://lkml.kernel.org/n/tip-zkjs8vstfnq9yq9r2jkh8kz0@git.kernel.org
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>